requirements.txt
requests==2.31.0
tqdm==4.67.1
pillow==12.0.0
numpy==2.2.6
pandas==2.3.3
sentence-transformers==2.2.2
scikit-learn==1.3.2
ollama==0.1.7
chromadb==0.4.22